requirements.txt
flask
logging
python-dotenv
requests
pymongo